Bring back default config for AllCops/Include
Behavior changed in 0.56.0 and we haven't noticed since Glue was still
running 0.52.0 and Duct-Tape wasn't running rubocop at all in its CI.

From https://docs.rubocop.org/en/stable/configuration

RuboCop comes with a comprehensive list of common ruby file names and
extensions. But, if you'd like RuboCop to check files that are not
included by default, you'll need to pass them in on the command line, or
to add entries for them under AllCops/Include. Remember that your
configuration files override RuboCops's defaults. In the following
example, we want to include foo.unusual_extension, but we also must copy
any other patterns we need from the overridden default.yml.

AllCops:
  Include:
    - foo.unusual_extension
    - '**/*.rb'
    - '**/*.gemfile'
    - '**/*.gemspec'
    - '**/*.rake'
    - '**/*.ru'
    - '**/Gemfile'
    - '**/Rakefile'

This behavior of Include (overriding default.yml) was introduced in
0.56.0 via #5882. This change allows people to include/exclude precisely
what they need to, without the defaults getting in the way.